+++ sdk/communication/communication-phone-numbers/recordings/node/siproutingclient__get_trunks/recording_can_retrieve_empty_trunks.js
@@ -1,0 +1,77 @@
+let nock = require('nock');
+
+module.exports.hash = "350ce45f52731288cad4f7db93a45b60";
+
+module.exports.testInfo = {"uniqueName":{},"newDate":{}}
+
+nock('https://endpoint', {"encodedQueryParams":true})
+  .get('/sip')
+  .query(true)
+  .reply(200, {"trunks":{"44.fqdn.com":{"sipSignalingPort":4567}},"routes":[{"description":"myFirstRoute's description","name":"myFirstRoute","numberPattern":"^+[1-9][0-9]{3,23}$","trunks":[]},{"description":"mySecondRoute's description","name":"mySecondRoute","numberPattern":"^+[1-9][0-9]{3,23}$","trunks":[]}]}, [
+  'Transfer-Encoding',
+  'chunked',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'MS-CV',
+  'jYnZock7dUabznefzt3MZw.0',
+  'Strict-Transport-Security',
+  'max-age=2592000',
+  'api-supported-versions',
+  '2021-05-01-preview, 2022-09-01-preview',
+  'X-Processing-Time',
+  '203ms',
+  'X-Cache',
+  'CONFIG_NOCACHE',
+  'X-Azure-Ref',
+  '0BKMDYwAAAAD8YNcEPsnpRYBRZMXeSzt+UFJHMDFFREdFMDkxOQBjYzkyNzU4ZC0wNWY3LTRhZDYtYWE1ZS0wZmE5NzE4ZDg5ODU=',
+  'Date',
+  'Mon, 22 Aug 2022 15:38:45 GMT'
+]);
+
+nock('https://endpoint', {"encodedQueryParams":true})
+  .patch('/sip', {"trunks":{"44.fqdn.com":null}})
+  .query(true)
+  .reply(200, {"trunks":{},"routes":[{"description":"myFirstRoute's description","name":"myFirstRoute","numberPattern":"^+[1-9][0-9]{3,23}$","trunks":[]},{"description":"mySecondRoute's description","name":"mySecondRoute","numberPattern":"^+[1-9][0-9]{3,23}$","trunks":[]}]}, [
+  'Transfer-Encoding',
+  'chunked',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'MS-CV',
+  'h0C/ZTYWkUWO7jIe8FrzGQ.0',
+  'Strict-Transport-Security',
+  'max-age=2592000',
+  'api-supported-versions',
+  '2021-05-01-preview, 2022-09-01-preview',
+  'X-Processing-Time',
+  '468ms',
+  'X-Cache',
+  'CONFIG_NOCACHE',
+  'X-Azure-Ref',
+  '0BaMDYwAAAACoTCQAktgfQq869CyySS6mUFJHMDFFREdFMDkxOQBjYzkyNzU4ZC0wNWY3LTRhZDYtYWE1ZS0wZmE5NzE4ZDg5ODU=',
+  'Date',
+  'Mon, 22 Aug 2022 15:38:45 GMT'
+]);
+
+nock('https://endpoint', {"encodedQueryParams":true})
+  .get('/sip')
+  .query(true)
+  .reply(200, {"trunks":{},"routes":[{"description":"myFirstRoute's description","name":"myFirstRoute","numberPattern":"^+[1-9][0-9]{3,23}$","trunks":[]},{"description":"mySecondRoute's description","name":"mySecondRoute","numberPattern":"^+[1-9][0-9]{3,23}$","trunks":[]}]}, [
+  'Transfer-Encoding',
+  'chunked',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'MS-CV',
+  'Tix7j+VcakaAhZ2lCBdxDA.0',
+  'Strict-Transport-Security',
+  'max-age=2592000',
+  'api-supported-versions',
+  '2021-05-01-preview, 2022-09-01-preview',
+  'X-Processing-Time',
+  '213ms',
+  'X-Cache',
+  'CONFIG_NOCACHE',
+  'X-Azure-Ref',
+  '0BaMDYwAAAADq9MdKt59WQJzx5SOOxpKTUFJHMDFFREdFMDkxOQBjYzkyNzU4ZC0wNWY3LTRhZDYtYWE1ZS0wZmE5NzE4ZDg5ODU=',
+  'Date',
+  'Mon, 22 Aug 2022 15:38:46 GMT'
+]);