+++ sdk/communication/communication-phone-numbers/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.2.0-beta.5 (Unreleased)
+## 1.2.0-alpha.20230303.1 (2023-03-03)
 
 ### Features Added
 
@@ -10,6 +10,16 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.2.0-alpha.20230303.1 (2023-03-03)
+- Added support for the API version 2023-01-01-preview.
+- Added ability to retrieve SIP trunk health properties.
+- Property "domains" added to sip configuration
+- Enabled property added to "trunk"
+
+## 1.2.0-alpha.20230221.1 (2023-02-21)
+- Added support for the API version 2023-01-01-preview.
+- Added ability to retrieve SIP trunk health properties.
 
 ## 1.2.0-beta.4 (2023-01-10)
 - Adds support for Azure Communication Services Phone Numbers Browse API Methods.